Add optional column sorting to handle ties
If sorting is done on a single column with more duplicates
than the page size it's possible than some rows are never
retrieved as we traverse through our datatable. This is because
of how order by together with limit and offset works in the database.

As a workaround for this problem we add a second column to sort by
in the case of ties.