fix bug with propagating datatables query to field
The problem was that I checked the 'field_name' attribute, which is
the name of the actual field on the model, instead of the name of the
filter.

This was forcing me to pick the names of the filters identical to the
field names, including full lookups, otherwise e.g. the regex flag
would not propagate.