implement ordering based on 'data' attribute only
This is what we need to get to the filter field, the name attribute is
useless for us, as the filter already defines the lookup.

Also the 'name' is meant to be just an identifier on the Datatable and
we don't need to add any magic property to it with a django-filter
based backend. That's what the filters are for, really.

See:

 - https://datatables.net/reference/option/columns.data
 - https://datatables.net/reference/option/columns.name
 - https://datatables.net/manual/server-side