factor out order_queryset method
to make it easier to override ordering completely.

This is necessary to be able to use asc and desc with nulls_first or
nulls_last, see
https://docs.djangoproject.com/en/2.2/ref/models/expressions/#django.db.models.Expression.asc

I need this in another project, but I won't have time to port the
solution to django-rest-framework-datatables in the near future.

If you want to override order_queryset and use asc() and desc() you'll
probably need to inline most of get_ordering and add those asc/desc
calls to the queryset appropriately.